# packages/api-client/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/staff/login": {
      "post": {
        "operationId": "StaffAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "mfa_required",
                        "mfa_enrollment_required"
                      ]
                    },
                    "mfaToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "mfaToken"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide (VALIDATION_ERROR) ou règle métier refusée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "423": {
            "description": "Compte verrouillé",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit atteinte (en-tête Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          }
        },
        "summary": "Première étape : courriel et mot de passe ; renvoie un jeton de passage pour le second facteur",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/staff/mfa/enroll": {
      "post": {
        "operationId": "StaffAuthController_enroll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mfaToken": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  }
                },
                "required": [
                  "mfaToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secret": {
                      "type": "string"
                    },
                    "otpauthUri": {
                      "type": "string"
                    },
                    "qrSvg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "secret",
                    "otpauthUri",
                    "qrSvg"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide (VALIDATION_ERROR) ou règle métier refusée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit atteinte (en-tête Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          }
        },
        "summary": "Inscription du second facteur (première connexion) : secret et QR pour une application d'authentification",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/staff/mfa/confirm": {
      "post": {
        "operationId": "StaffAuthController_confirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mfaToken": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "mfaToken",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "accessToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "created": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "phone": {
                          "type": "string",
                          "pattern": "^\\+[1-9]\\d{6,14}$"
                        },
                        "email": {
                          "nullable": true,
                          "type": "string",
                          "format": "email",
                          "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                        },
                        "firstName": {
                          "nullable": true,
                          "type": "string"
                        },
                        "lastName": {
                          "nullable": true,
                          "type": "string"
                        },
                        "language": {
                          "type": "string",
                          "enum": [
                            "fr",
                            "en"
                          ]
                        },
                        "primaryRole": {
                          "type": "string",
                          "enum": [
                            "client",
                            "driver",
                            "partner",
                            "investor",
                            "operator",
                            "admin",
                            "agent",
                            "finance",
                            "readonly"
                          ]
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "client",
                              "driver",
                              "partner",
                              "investor",
                              "operator",
                              "admin",
                              "agent",
                              "finance",
                              "readonly"
                            ]
                          }
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active",
                            "blocked",
                            "deleted"
                          ]
                        },
                        "termsAcceptedAt": {
                          "nullable": true,
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                        },
                        "privacyPolicyVersion": {
                          "nullable": true,
                          "type": "string"
                        },
                        "privacyPolicyCurrentVersion": {
                          "type": "string"
                        },
                        "privacyPolicyAccepted": {
                          "type": "boolean"
                        },
                        "linkedProviders": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "apple",
                              "google"
                            ]
                          }
                        },
                        "mfaEnabled": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                        }
                      },
                      "required": [
                        "id",
                        "phone",
                        "email",
                        "firstName",
                        "lastName",
                        "language",
                        "primaryRole",
                        "roles",
                        "status",
                        "termsAcceptedAt",
                        "privacyPolicyVersion",
                        "privacyPolicyCurrentVersion",
                        "privacyPolicyAccepted",
                        "linkedProviders",
                        "mfaEnabled",
                        "createdAt"
                      ]
                    },
                    "backupCodes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "tokenType",
                    "accessToken",
                    "expiresIn",
                    "refreshToken",
                    "created",
                    "user",
                    "backupCodes"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide (VALIDATION_ERROR) ou règle métier refusée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit atteinte (en-tête Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {},
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ]
                }
              }
            }
          }
        },
        "summary": "Confirme l'inscription avec un premier code : codes de secours (affichés une seule fois) et jetons",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Neomoov API",
    "description": "API de la plateforme Neomoov (Groupe NSK Inc.). Préfixe /v1.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}